services/labs/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
email-validator
python-jose
passlib[bcrypt]
databases
python-multipart
psycopg2-binary
python-dotenv
